requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.6.1
pandas==2.1.4
sqlalchemy==2.0.23
python-multipart==0.0.6
aiohttp==3.9.1
sentence-transformers==2.2.2
uvicorn[standard]==0.24.0